tech(base): add react router
