+++ package.json
@@ -26,7 +26,8 @@
     "react": "^17.0.2",
     "react-camouflage": "^1.3.13",
     "react-dom": "^17.0.2",
-    "react-i18next": "^11.15.3"
+    "react-i18next": "^11.15.3",
+    "react-router-dom": "^6.3.0"
   },
   "devDependencies": {
     "@babel/core": "^7.16.12",
