+++ src/modules/Profile/Profile.js
@@ -1,0 +1,15 @@
+import React from 'react'
+
+const Profile = () => {
+
+  return (
+    <section>
+      <div>
+        <span></span>
+        <h1>Profile Page here</h1>
+      </div>
+    </section>
+  )
+}
+
+export default Profile
