+++ src/utils/route-paths.js
@@ -1,0 +1,5 @@
+export default {
+    Home: '/',
+    Profile: '/profile',
+  }
+
